Added fix for Chinese characters in Micromeritics datetime cell
For at least one user Micromeritics xls will be saved with the Chinese characters "下午" in the experiment time cell. I believe this represents something like 'PM'. Because exact accuracy in the time is currently not necessary and for simplicity this fix will remove these characters which then allows the date time parser to continue without errors.